# Reduce memory footprint of `dss_line`