--- a/meegkit/dss.py
+++ b/meegkit/dss.py
@@ -197,24 +197,21 @@ def dss_line(X, fline, sfreq, nremove=1, nfft=1024, nkeep=None, blocksize=None,
         blocksize = n_samples
 
     # Recentre data
-    X = demean(X)
+    X = demean(X, inplace=True)
 
     # Cancel line_frequency and harmonics + light lowpass
     X_filt = smooth(X, sfreq / fline)
-
-    # Subtract clean data from original data. The result is the artifact plus
-    # some residual biological signal
-    X_noise = X - X_filt
-
+  
+    # X - X_filt results in the artifact plus some residual biological signal
     # Reduce dimensionality to avoid overfitting
     if nkeep is not None:
-        cov_X_res = tscov(X_noise)[0]
+        cov_X_res = tscov(X - X_filt)[0]
         V, _ = pca(cov_X_res, nkeep)
-        X_noise_pca = X_noise @ V
+        X_noise_pca = (X - X_filt) @ V
     else:
-        X_noise_pca = X_noise.copy()
+        X_noise_pca = (X - X_filt).copy()
         nkeep = n_chans
-
+        
     # Compute blockwise covariances of raw and biased data
     n_harm = np.floor((sfreq / 2) / fline).astype(int)
     c0 = np.zeros((nkeep, nkeep))
@@ -226,10 +223,9 @@ def dss_line(X, fline, sfreq, nremove=1, nfft=1024, nkeep=None, blocksize=None,
             X_block = X_block.transpose(1, 2, 0)
 
         # bias data
-        X_bias = gaussfilt(X_block, sfreq, fline, fwhm=1, n_harm=n_harm)
         c0 += tscov(X_block)[0]
-        c1 += tscov(X_bias)[0]
-
+        c1 += tscov(gaussfilt(X_block, sfreq, fline, fwhm=1, n_harm=n_harm))[0]
+    
     # DSS to isolate line components from residual
     todss, _, pwr0, pwr1 = dss0(c0, c1)
 
@@ -244,16 +240,15 @@ def dss_line(X, fline, sfreq, nremove=1, nfft=1024, nkeep=None, blocksize=None,
     # Remove line components from X_noise
     idx_remove = np.arange(nremove)
     X_artifact = matmul3d(X_noise_pca, todss[:, idx_remove])
-    X_res = tsr(X_noise, X_artifact)[0]  # project them out
-
+    X_res = tsr(X - X_filt, X_artifact)[0]  # project them out
     # reconstruct clean signal
     y = X_filt + X_res
-    artifact = X - y
 
     # Power of components
     p = wpwr(X - y)[0] / wpwr(X)[0]
     print('Power of components removed by DSS: {:.2f}'.format(p))
-    return y, artifact
+    # return the reconstructed clean signal, and the artifact
+    return y, X - y
 
 
 def dss_line_iter(data, fline, sfreq, win_sz=10, spot_sz=2.5,