--- a/meegkit/tspca.py
+++ b/meegkit/tspca.py
@@ -107,7 +107,6 @@ def tsr(X, R, shifts=None, wX=None, wR=None, keep=None, thresh=1e-12):
         Weights applied by TSR.
 
     """
-    from psutil import Process
     ndims = X.ndim
     X = unsqueeze(X)
     R = unsqueeze(R)
@@ -184,7 +183,7 @@ def tsr(X, R, shifts=None, wX=None, wR=None, keep=None, thresh=1e-12):
     for t in np.arange(n_trials_X):
         r = multishift(R[..., t], shifts, reshape=True)
         y[..., t] = X[:z.shape[0], :, t] - (r @ regression)
-    
+
     y, mean2 = demean(y, wX, return_mean=True, inplace=True)
     idx = np.arange(offset1, initial_samples - offset2)
     mean_total = mean1 + mean2