--- a/meegkit/utils/sig.py
+++ b/meegkit/utils/sig.py
@@ -345,13 +345,13 @@ def gaussfilt(data, srate, f, fwhm, n_harm=1, shift=0, return_empvals=False,
         fx = fx + gauss
 
     # filter
-    tmp = np.fft.fft(data, axis=0)
     if data.ndim == 2:
-        tmp *= fx[:, None]
+        filtdat = 2 * np.real(np.fft.ifft(
+                              np.fft.fft(data, axis=0)*fx[:, None], axis=0))
     elif data.ndim == 3:
-        tmp *= fx[:, None, None]
-
-    filtdat = 2 * np.real(np.fft.ifft(tmp, axis=0))
+        filtdat = 2 * np.real(np.fft.ifft(
+                              np.fft.fft(data, axis=0)*fx[:, None, None],
+                              axis=0))
 
     if return_empvals or show:
         empVals[0] = hz[idx_p]