@nbara 2021-12-02
Thanks @eort. I have looked at the changes only quickly, but I can already tell you that the CI do not pass any more (you can check that locally by running `make pep` from meegkit's root directory)

I think only a subset of those changes are beneficial (the "inplace" recentring for instance). Some others I am less sure about (for instance when you compute `X-X_filt` 4 times in the `dss_line()` code). 

I will try to post proper benchmarks (computation time and memory consumption) soon to get to the bottom of this.