@eort 2021-12-02
Yeah I wasn't sure about those ones either, but for me saving memory was more critical than saving time. In any case, looking forward to those benchmarks!

(Hope CI is happier now)