@nbara 2021-12-02
> (Hope CI is happier now)

Bah, I meant run `pytest --noplots` to run the unit tests, which are broken (due to the changes in demean, apparently). 

I'll post the benchmarks shortly